Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main(){
-    int a,b,c,x;
-    scanf("%d%d%d",&a,&b,&c);
+    float a,b,c,x;
+    scanf("%f%f%f",&a,&b,&c);
     x=(a+b+c)/3; 
-    printf("Average: %d",x);
+    printf("Average: %.2f",x);
     return 0;
 }
